backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.0.3
#httpx==0.25.2
python-dotenv==1.0.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
pillow==10.1.0
openai==1.6.1
aiofiles==23.2.1


